refactor: PageStyle 이름 수정
